pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ordinal-classifier"
version = "0.1.0"
description = "Ordinal classifier using ResNet-*"
authors = [
    {name = "Allen Day", email = "allenday@users.github.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "fastai>=2.7.0",
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "pillow>=9.0.0",
    "pandas>=1.5.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "click>=8.0.0",
    "pathlib",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]

[project.scripts]
ordinal-classifier = "ordinal_classifier.cli:main"

[project.urls]
Homepage = "https://github.com/allenday/ordinal-classifier"
Repository = "https://github.com/allenday/ordinal-classifier"
Documentation = "https://github.com/allenday/ordinal-classifier/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["ordinal_classifier*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88 

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=ordinal_classifier --cov-report=term-missing --cov-report=html -m 'not integration'"
markers = [
    "integration: marks tests as integration tests (deselected by default)",
]
